README: Add libvirt instructions
libvirt 5.0.0 supports ARMv6 guests and, as of commit
e362b8bbd49f, the kernel built using the scripts in this
repository can drive all the virtual hardware in such a guest.

Provide some instructions on how to use libvirt instead of
spawning QEMU directly.

Note that, while VirtIO RNG support is not yet available in any
of the pre-built kernel images, the corresponding configuration
toggles have been enabled in commit 4938f7a626b5, so the next
time a kernel image is built they will be picked up. Given that,
it makes sense to start adding the device to guests right away.