tools: Refresh linux-arm.patch
Just like the unversioned configuration file is used
unless a more specific version is requested, and in particular
every time the latest kernel is built, it makes sense that
the patch would have the same behavior.

This avoid the awkward situation where, as it happened when
introducing 4.14.79, a new file is introduced despite the
previous version of the patch being applicable as-is.

Going forward, creating a named patch file will only be
necessary when adding support for a new major release of
Raspbian.